_triesArray is now always sliced prior to being included in _queries (#9)
Removed conditional slice in .query that would never be called and replaced it with a slice that is always
called otherwise a reference to this._triesArray was being included in the _queries array and subsequently
being destroyed by the normal operation of _ontimeout.   This resulted in the emptying of this._triesArray
and failure of subsequent queries.